Check state of GPIO line at boot time for vccaux workaround requirement
With most boot modes, as soon as the board is powered on, Versal starts to boot.  PLM asserts GPIO line to tell SC that it is time to apply the workaround.  In such a mode, SC is also booting and it will miss the 0 to 1 transition on GPIO line, therefore, no workaround is applied.  This commit fixes that issue.

Signed-off-by: Saeed Nowshadi <saeed.nowshadi@xilinx.com>