Add support for customization of commands
Constraints could be defined for any sc_app command.  The constraints could
be defined as additional external or internal operations to be performed prior
to actual execution of sc_app command.  The external operations are performed
by invoking external scripts/binaries, and the internal operations are performed
by invoking internally defined routines.  Currently, these additional operations
can be defined for pre-execution of actual command, but that could be expanded
for post-execution as well.  Additionally, new custom commands could be defined
to be invoked through sc_app interface but passed to external scripts/binaries
for execution.

Signed-off-by: Saeed Nowshadi <saeed.nowshadi@xilinx.com>
Acked-by: Jesus De Haro <jesus.de-haro@xilinx.com>