Create a thread to manage fancontrol process
Versal's temperature is available only when SysMon's I2C interface is configured
by a PDI.  If there is no PDI loaded, the temperature won't be available and the
running fancontrol process terminates.  The fancontrol thread is responsible for
periodically checking for a valid temperature, and upon detection restart the
fancontrol process.

Signed-off-by: Saeed Nowshadi <saeed.nowshadi@amd.com>