Better handling of Unknown board type
Remove the startup script as it is no longer needed and it blocks
sc_appd from starting if we have an Unknown board.

If the board type is Unknown, meaning we don't have a matching JSON file,
then we likely don't have the firmware package either, so skip over any
jtag operations.

Add a script that allows us to install the board support files

Signed-off-by: Jonathan Stroud <jonathan.stroud@amd.com>
Signed-off-by: Saeed Nowshadi <saeed.nowshadi@amd.com>